removes publishing responsibility from commit store
This is important because in our new architecture
the commit saving and publishing will be completely
separated. The publisher will observe the mongodb
for new commits and does not know about the
commit store. This also means that communication
between apps etc. is more flexible because any
process can insert commits as long as the format
is right.