findAndModify commit when timing out, to eliminate race condition
findAndModify is now used with a query that excludes commits that may
have since finished processing, so that the timeout can be ignored.
This is the only real inconsistency we need to guard for as if the
process finishes after the failure (or is retried), the commit gets
updated as processed, so is representing the true state of the commit.
As a bonus this approach also reduces the database interaction from 2
to 1 call